+++ src/tokenizer.rs
@@ -578,8 +578,23 @@ impl<'a> fmt::Display for StringLiteralQuoteType {
 
 lazy_static! {
     static ref PATTERN_IDENTIFIER: Regex = Regex::new(r"[^\W\d]+\w*").unwrap();
-    static ref PATTERN_NUMBER: Regex =
-        Regex::new(r"^((0x[A-Fa-f\d]+)|(((\d*\.\d+)|(\d+))([eE]-?\d+)?))").unwrap();
+    static ref PATTERN_NUMBER: Regex = {
+        let mut set = Vec::new();
+
+        // Basic numbers (123456.346e123)
+        set.push(r"(((\d*\.\d+)|(\d+))([eE]-?\d+)?)");
+
+        // Hexadecimal
+        set.push(r"(0x[A-Fa-f\d]+)");
+
+        if cfg!(feature = "roblox") {
+            // Binary literals
+            set.push(r"(0b[01]+)");
+        }
+
+        Regex::new(&format!("^({})", set.join("|"))).unwrap()
+    };
+
     static ref PATTERN_COMMENT_MULTI_LINE_BEGIN: Regex = Regex::new(r"--\[(=*)\[").unwrap();
     static ref PATTERN_COMMENT_SINGLE_LINE: Regex = Regex::new(r"--([^\n]*)").unwrap();
     static ref PATTERN_STRING_MULTI_LINE_BEGIN: Regex = Regex::new(r"\[(=*)\[").unwrap();